fix: update scrape references method, refactor
